feat(components): Implement Phase 1 of the systemic development plan
This commit implements the first phase of the systemic development plan, focusing on critical bug fixes and API standardization across several components.

The following changes have been made:

- **Button:** The `typepe` property has been corrected to `r#type`.
- **Checkbox:** The `validitype` property has been corrected to `validity`.
- **Fab:**
    - The `kind` property has been renamed to `variant` for consistency.
    - The `fab_type` property is now used to distinguish between `md-fab` and `md-branded-fab`.
    - The `variant` and `size` properties now use type-safe enums (`FabVariant`, `FabSize`) instead of strings.
- **CircularProgress:** The `value` and `max` properties have been changed from `usize` to `f32` to support fractional values.

All usage examples in the `usages` and `matdemo` crates have been updated to reflect these API changes, and all build warnings have been resolved.

+++ src/circular_progress.rs
@@ -3,10 +3,10 @@ use yew::prelude::*;
 pub struct Props {
     /// Progress to display, a fraction between 0 and <code>max</code>.
     #[prop_or_default]
-    pub value: usize,
+    pub value: f32,
     /// Maximum progress to display, defaults to 1.
-    #[prop_or(1)]
-    pub max: usize,
+    #[prop_or(1.0)]
+    pub max: f32,
     /// Whether or not to display indeterminate progress, which gives no indication to how long an
     /// activity will take.
     #[prop_or_default]
